[BACKEND] Refactor http responses classess
