--- app/backend/src/main/java/pl/lodz/p/it/inz/sgruda/multiStore/mok/payloads/response/UserIdentityAvailability.java
@@ -1,14 +1,0 @@
-package pl.lodz.p.it.inz.sgruda.multiStore.mok.payloads.response;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class UserIdentityAvailability {
-    private Boolean available;
-
-    public UserIdentityAvailability(Boolean available) {
-        this.available = available;
-    }
-}
