--- app/backend/src/main/java/pl/lodz/p/it/inz/sgruda/multiStore/mok/payloads/response/UserSummary.java
@@ -1,18 +1,0 @@
-package pl.lodz.p.it.inz.sgruda.multiStore.mok.payloads.response;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class UserSummary {
-    private Long id;
-    private String username;
-    private String name;
-
-    public UserSummary(Long id, String username, String name) {
-        this.id = id;
-        this.username = username;
-        this.name = name;
-    }
-}
