+++ app/backend/src/main/java/pl/lodz/p/it/inz/sgruda/multiStore/security/RestAuthenticationEntryPoint.java
@@ -4,20 +4,15 @@ package pl.lodz.p.it.inz.sgruda.multiStore.security;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.extern.java.Log;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.http.server.ServletServerHttpResponse;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.stereotype.Component;
-import pl.lodz.p.it.inz.sgruda.multiStore.mok.payloads.response.ApiResponse;
+import pl.lodz.p.it.inz.sgruda.multiStore.responses.ApiResponse;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.io.OutputStream;
 
 @Log
 @Component
